kubermatic-installer: deploy ingress-controller or envoy gateway controller in separate seed
